requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchmetrics>=0.11.0
pandas>=2.0.3
numpy>=1.24.3
opencv-python-headless>=4.7.0
Pillow>=9.5.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
gunicorn>=20.1.0
gdown>=4.7.1